refactor : navigation info 분리 및 bottom bar 컴포넌트로 이름 변경 #49
