+++ src/App.tsx
@@ -1,6 +1,8 @@
 import { useRoutes } from 'react-router-dom';
 
-import NavigationBar from '@components/NavigationBar';
+import BottomBar from '@components/BottomBar';
+import Login from '@components/Login';
+import Search from '@components/Search';
 import { Mobile } from '@query/mediaQuery';
 import { routes } from '@router';
 import Styles from '@styles';
@@ -12,8 +14,10 @@ const App = () => {
     <div className='App'>
       <Styles>
         {element}
+        <Login />
+        <Search />
         <Mobile>
-          <NavigationBar />
+          <BottomBar />
         </Mobile>
       </Styles>
     </div>
